requirements.txt
python-jose
fastapi[standard]~=0.122
sqlalchemy~=2.0.44
passlib~=1.7.4
dotenv~=0.9.9
python-dotenv~=1.2.1
pydantic~=2.12.4